Added birth and survive rules to life parameters
Added the parsing of strings that represent life-like rules in the form
"B...S..." where the number of neighbors can be listed after the 'B' and
'S' to indicate when a cell should be born or when an alive cell should
survive. An argument was added to allow this to be passed on the command
line. Updated the life implementation to use the birth and survive rules
to determine the generation. Personal favorite life-like rule at the
moment: "B3S134"