Parallelized computation of next generation of cells
Created worker threads that compute the next generation of cells given a
chunk of the current grid. The results are fed back through a channel
for the main thread to modify the next generation grid. When all results
are in, the next and current are swapped (this is the same as before).
Added additional trait bounds required to send cell (and their
parameter) types across thread boundaries. This did not break
compatibility with existing cell types.